+++ apps/simulation-study/Main.hs
@@ -1,19 +1,22 @@
 {-# LANGUAGE OverloadedStrings #-}
+{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE RecordWildCards #-}
 
 module Main where
 
+import BDSCOD.Conditioning
 import BDSCOD.Llhd
 import BDSCOD.Types
 import BDSCOD.Utility
-import BDSCOD.Conditioning
-import Epidemic.Types
+import qualified Data.Aeson as Json
 import qualified Data.ByteString.Lazy as L
-import Data.Csv
+import qualified Data.Csv as Csv
 import Data.List (intercalate)
 import Data.Maybe
 import qualified Epidemic.BDSCOD as SimBDSCOD
+import Epidemic.Types
 import qualified Epidemic.Utility as SimUtil
+import GHC.Generics
 
 
 condLlhdAndNB :: [Observation] -> Parameters -> Time -> Bool -> (LogLikelihood, NegativeBinomial)
@@ -32,7 +35,7 @@ llhdsWriteFile fp d ps duration conditionLlhd =
       let x = condLlhdAndNB d p duration conditionLlhd
       appendFile fp $ output p x
       llhdsWriteFile fp d ps' duration conditionLlhd
-      where output (x1, x2, x3, [(_, x4)], x5, [(_, x6)]) (x7, x8) =
+      where output (x1, x2, x3, ((_, x4):_), x5, [(_, x6)]) (x7, x8) =
               intercalate "," $
               map show [x1, x2, x3, x4, x5, x6, x7] ++ [show x8 ++ "\n"]
 
@@ -56,29 +59,40 @@ linspace x1 x2 n = [x1 + fromIntegral i * delta | i <- [0 .. (n - 1)]]
   where
     delta = (x2 - x1) / (fromIntegral n - 1)
 
+data SimStudyParams =
+  SimStudyParams
+    { outputEventsFile :: FilePath
+    , outputObservationsFile :: FilePath
+    , outputLlhdFile :: FilePath
+    , simDuration :: Time
+    , simLambda :: Rate
+    , simMu :: Rate
+    , simPsi :: Rate
+    , simRho :: Probability
+    , simRhoTimes :: [Time]
+    , simOmega :: Rate
+    , simNu :: Probability
+    , simNuTime :: Time
+    }
+  deriving (Show, Generic)
+
+instance Json.FromJSON SimStudyParams
+
+readConfigFile :: FilePath -> IO (Maybe SimStudyParams)
+readConfigFile fp = Json.decode <$> L.readFile fp
+
 main :: IO ()
-main =
-  let outputFileSimulation1 = "out/simulated-events.txt"
-      outputFileSimulation2 = "out/simulated-all-events.txt"
-      outputFileObservations = "out/simulated-events-observed.txt"
-      outputFileLlhdValues = "out/simulation-study-llhds.csv"
-      simDuration = 6.0
-      simLambda = 1.5
-      simMu = 0.3
-      simPsi = 0.3
-      simRho = 0.15
-      simRhoTime = 3.0
-      simOmega = 0.3
-      simNu = 0.15
-      simNuTime = 4.0
-      simParams = (simLambda, simMu, simPsi, [(simRhoTime,simRho)], simOmega, [(simNuTime,simNu)])
+main = do
+  config <- readConfigFile "out/config.json"
+  let SimStudyParams{..} = fromJust config
+      simParams = (simLambda, simMu, simPsi, [(rt,simRho) | rt <- simRhoTimes], simOmega, [(simNuTime,simNu)])
       infParamss =
-        [(l, simMu, simPsi, [(simRhoTime,simRho)], simOmega, [(simNuTime,simNu)]) | l <- linspace 1.0 8.0 200] ++
-        [(simLambda, m, simPsi, [(simRhoTime,simRho)], simOmega, [(simNuTime,simNu)]) | m <- linspace 0.01 2.0 200] ++
-        [(simLambda, simMu, p, [(simRhoTime,simRho)], simOmega, [(simNuTime,simNu)]) | p <- linspace 0.01 2.0 200] ++
-        [(simLambda, simMu, simPsi, [(simRhoTime,r)], simOmega, [(simNuTime,simNu)]) | r <- linspace 0.01 0.40 400] ++
-        [(simLambda, simMu, simPsi, [(simRhoTime,simRho)], o, [(simNuTime,simNu)]) | o <- linspace 0.01 2.0 200] ++
-        [(simLambda, simMu, simPsi, [(simRhoTime,simRho)], simOmega, [(simNuTime,n)]) | n <- linspace 0.01 0.40 400]
+        [(l, simMu, simPsi, [(rt,simRho) | rt <- simRhoTimes], simOmega, [(simNuTime,simNu)]) | l <- linspace 1.0 8.0 200] ++
+        [(simLambda, m, simPsi, [(rt,simRho) | rt <- simRhoTimes], simOmega, [(simNuTime,simNu)]) | m <- linspace 0.01 2.0 200] ++
+        [(simLambda, simMu, p, [(rt,simRho) | rt <- simRhoTimes], simOmega, [(simNuTime,simNu)]) | p <- linspace 0.01 2.0 200] ++
+        [(simLambda, simMu, simPsi, [(rt,r) | rt <- simRhoTimes], simOmega, [(simNuTime,simNu)]) | r <- linspace 0.01 0.40 400] ++
+        [(simLambda, simMu, simPsi, [(rt,simRho) | rt <- simRhoTimes], o, [(simNuTime,simNu)]) | o <- linspace 0.01 2.0 200] ++
+        [(simLambda, simMu, simPsi, [(rt,simRho) | rt <- simRhoTimes], simOmega, [(simNuTime,n)]) | n <- linspace 0.01 0.40 400]
       simConfig = SimBDSCOD.configuration simDuration simParams
       conditionUponObservation = True
    in if isNothing simConfig
@@ -87,14 +101,13 @@ main =
       else
         do putStrLn appMessage
            simEvents <- SimUtil.simulation True (fromJust simConfig) SimBDSCOD.allEvents
-           Prelude.writeFile outputFileSimulation1 $ intercalate "\n" (map show simEvents)
-           L.writeFile outputFileSimulation2 $ encode simEvents
+           Prelude.writeFile outputEventsFile $ intercalate "\n" (map show simEvents)
            let obs =
                  eventsAsObservations $
                  SimBDSCOD.observedEvents simEvents
                numSimEvents = length simEvents
                numObs = length obs
-           Prelude.writeFile outputFileObservations $ intercalate "\n" (map show obs)
+           Prelude.writeFile outputObservationsFile $ intercalate "\n" (map show obs)
            putStrLn $ "Number of events in the simulation: " ++ show numSimEvents
            putStrLn $ "Number of events in the dataset: " ++ show numObs
-           llhdsWriteFile outputFileLlhdValues obs infParamss simDuration conditionUponObservation
+           llhdsWriteFile outputLlhdFile obs infParamss simDuration conditionUponObservation